Add files via upload
Converted the Python script into a Jupyter Notebook.
Adapted the code to match the syntax of the latest version of Python.

Note: the files "skynet_network.txt" and "skynet_pred_out" have to be created in the working directory to function. The next lines to be fixed are in cell 2 line 208 and line 235 as they don't find the needed file (SkyNet and CalPred).